+++ R/pkgSnapshot.R
@@ -25,7 +25,7 @@ pkgSnapshot <- function(packageVersionFile = "packageVersions.txt", libPaths, st
   origLibPaths <- setLibPaths(libPaths, standAlone)
   on.exit({.libPaths(origLibPaths)}, add = TRUE)
 
-  ip <- as.data.table(.installed.pkgs(lib.loc = libPaths))
+  ip <- as.data.table(.installed.pkgs(lib.loc = libPaths, which = character()))
   instPkgs <- ip$Package
   instVers <- ip$Version
   names(instPkgs) <- instPkgs